refactor: Move GroupByAndAggregate into service/expr.rs
